WebCL-conformance/conformance/extension/WEBCL_html_image/cl_commandQueue_enqueueWriteBuffer_with_imageData.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueWriteBuffer with imageData conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<div style="position:absolute; left:0px; top:0px; visibility:hidden" id="x">
    <canvas id="inputCanvas" width = 200 height = 100/>
</div>
<script>
try {
    debug("WebCLCommandQueue :: enqueueWriteBuffer() with imageData test case.");
    debug("( WebCLBuffer buffer, CLboolean blockingWrite, CLuint bufferOffset, ImageData srcPixels, optional sequence<WebCLEvent>? eventWaitList, optional WebCLEvent? event)\n");

    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var canvas = document.getElementById("inputCanvas");
    var canvasContext = canvas.getContext("2d");
    var imgData = canvasContext.createImageData(100, 100);
    for (var i = 0; i < imgData.data.length; i+=4) {
        imgData.data[i+0] = 255;
        imgData.data[i+1] = 0;
        imgData.data[i+2] = 0;
        imgData.data[i+3] = 255;
    }
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var bufferOffset = 0;

    var bufferSize = imgData.width * imgData.height * 4;
    var webCLBuffer = wtu.createBuffer(webCLContext, webcl.MEM_READ_ONLY, bufferSize);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(webCLBuffer, true, bufferOffset, imgData)", "WEBCL_EXTENSION_NOT_ENABLED");
    if (!wtu.enableExtension(webcl, "WEBCL_html_image")) {
        throw { description : "Failed to enable extension : WEBCL_html_image. Exitting..." };
    }

    wtu.releaseAll(webCLContext);
    webCLContext = wtu.createContext();
    webCLCommandQueue = wtu.createCommandQueue(webCLContext);

    // MEM_READ_ONLY Buffers.
    var webCLBufferRead = wtu.createBuffer(webCLContext, webcl.MEM_READ_ONLY, bufferSize);
    shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferRead, true, bufferOffset, imgData)");
    shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferRead, false, bufferOffset, imgData)");

    // MEM_WRITE_ONLY Buffers.
    var webCLBufferWrite = wtu.createBuffer(webCLContext, webcl.MEM_WRITE_ONLY, bufferSize);
    shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferWrite, true, bufferOffset, imgData)");
    shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferWrite, false, bufferOffset, imgData)");

    // MEM_READ_WRITE Buffers.
    var webCLBufferReadWrite = wtu.createBuffer(webCLContext, webcl.MEM_READ_WRITE, bufferSize);
    shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, imgData)");
    shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, false, bufferOffset, imgData)");

    // Testing for nullable parameters.
    shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, imgData, [])")
    shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, imgData, [], null)")
    shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, imgData, null)")
    shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, imgData, null, null)")
    webCLEvent = wtu.createEvent();
    shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, imgData, null, webCLEvent)")
    webCLEvent = wtu.createEvent();
    shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, imgData, [], webCLEvent)")

    bufferOffsetNonZero = 10;
    var bufferSizeBig = (imgData.width * imgData.height * 4) + bufferOffsetNonZero;
    var webCLBufferBig = wtu.createBuffer(webCLContext, webcl.MEM_READ_WRITE, bufferSizeBig);
    shouldBeUndefined("webCLCommandQueue.enqueueWriteBuffer(webCLBufferBig, true, bufferOffsetNonZero, imgData)");

    // Testing for exception.
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffsetNonZero, imgData)", "INVALID_VALUE");

    var releasedBuffer = wtu.createBuffer(webCLContext, webcl.MEM_READ_WRITE, bufferSize);
    wtu.release(releasedBuffer);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(releasedBuffer, true, bufferOffset, imgData)", "INVALID_MEM_OBJECT");
    var imgDataInvalid = canvasContext.createImageData(-10, -10);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, canvas)", "INVALID_HOST_PTR");

    var bufferOffsetInvalid = bufferSize + 1;
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffsetInvalid, imgData)", "INVALID_VALUE");

    var canvasContext = canvas.getContext("2d");
    var imageData = canvasContext.createImageData(10, 10);
    var bufferSizeSmall = imgData.data.length - 1;
    var webCLBufferSmall = wtu.createBuffer(webCLContext, webcl.MEM_READ_WRITE, bufferSizeSmall);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(webCLBufferSmall, true, bufferOffset, imgData)", "INVALID_VALUE");

    // Testing for exception "INVALID_CONTEXT"
    var webCLPlatform = wtu.getPlatform();
    var devices = wtu.getDevices(webCLPlatform, webcl.DEVICE_TYPE_ALL);
    if (devices.length > 1) {
        var webCLDevice1 = devices[0];
        var webCLContext1 = wtu.createContext(webCLDevice1);
        var webCLCommandQueue1 = wtu.createCommandQueue(webCLContext1, webCLDevice1);
        var webCLBuffer1 = wtu.createBuffer(webCLContext1, webcl.MEM_READ_WRITE, bufferSize);

        var webCLDevice2 = devices[1];
        var webCLContext2 = wtu.createContext(webCLDevice2);
        var webCLCommandQueue2 = wtu.createCommandQueue(webCLContext2, webCLDevice2);
        var webCLBuffer2 = wtu.createBuffer(webCLContext2, webcl.MEM_READ_WRITE, bufferSize);

        shouldThrowExceptionName("webCLCommandQueue1.enqueueWriteBuffer(webCLBuffer2, true, bufferOffset, imgData)", "INVALID_CONTEXT");
        shouldThrowExceptionName("webCLCommandQueue2.enqueueWriteBuffer(webCLBuffer1, true, bufferOffset, imgData)", "INVALID_CONTEXT");

        var webCLUserEvent1 = wtu.createUserEvent(webCLContext1);
        var webCLUserEvent2 = wtu.createUserEvent(webCLContext2);
        shouldThrowExceptionName("webCLCommandQueue1.enqueueWriteBuffer(webCLBuffer1, false, bufferOffset, imgData, [webCLUserEvent2])", "INVALID_CONTEXT");
        shouldThrowExceptionName("webCLCommandQueue2.enqueueWriteBuffer(webCLBuffer2, false, bufferOffset, imgData, [webCLUserEvent1])", "INVALID_CONTEXT");
        wtu.setStatus(webCLUserEvent1, webcl.COMPLETE);
        wtu.setStatus(webCLUserEvent2, webcl.COMPLETE);

        var webCLEvent1 = wtu.createEvent();
        webCLCommandQueue1.enqueueWriteBuffer(webCLBuffer1, true, bufferOffset, imgData, [], webCLEvent1);
        shouldThrowExceptionName("webCLCommandQueue2.enqueueWriteBuffer(webCLBuffer2, true, bufferOffset, imgData, [webCLEvent1])", "INVALID_CONTEXT");
        var webCLEvent2 = wtu.createEvent();
        webCLCommandQueue2.enqueueWriteBuffer(webCLBuffer2, true, bufferOffset, imgData, [], webCLEvent2);
        shouldThrowExceptionName("webCLCommandQueue1.enqueueWriteBuffer(webCLBuffer1, true, bufferOffset, imgData, [webCLEvent2])", "INVALID_CONTEXT");
    }

    var releasedEvent = wtu.createUserEvent(webCLContext);
    wtu.release(releasedEvent);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, imgData, [releasedEvent])", "INVALID_EVENT_WAIT_LIST");

    var webCLUserEvent = wtu.createUserEvent(webCLContext);
    wtu.setStatus(webCLUserEvent, -1);
    shouldThrowExceptionName("webCLCommandQueue.enqueueWriteBuffer(webCLBufferReadWrite, true, bufferOffset, imgData, [webCLUserEvent])", "EXEC_STATUS_ERROR_FOR_EVENTS_IN_WAIT_LIST");

} catch (e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
